Move approve/decline tour logic to admin service
