+++ SelfGuidedTours.Core/Contracts/IAdminService.cs
@@ -7,5 +7,8 @@ namespace SelfGuidedTours.Core.Contracts
     public interface IAdminService
     {
         Task<IEnumerable<AllToursToAdminDTO>> GetAllToursAsync(Status status);
+        Task<ApiResponse> ApproveTourAsync(int id);
+
+        Task<ApiResponse> RejectTourAsync(int id);
     }
 }
