+++ SelfGuidedTours.Core/Contracts/ITourService.cs
@@ -18,8 +18,5 @@ namespace SelfGuidedTours.Core.Contracts
 
         TourResponseDto MapTourToTourResponseDto(Tour tour);
 
-        Task<ApiResponse> ApproveTourAsync(int id);
-
-        Task<ApiResponse> RejectTourAsync(int id);
     }
 }
