+++ SelfGuidedTours.Core/Services/TourService.cs
@@ -27,24 +27,6 @@ namespace SelfGuidedTours.Core.Services
             this.landmarkService = landmarkService;
         }
 
-        public async Task<ApiResponse> ApproveTourAsync(int id)
-        {
-            var tour = await repository.GetByIdAsync<Tour>(id)
-                ?? throw new KeyNotFoundException(TourNotFoundErrorMessage);
-
-            if (tour.Status == Status.Approved) throw new InvalidOperationException(TourAlreadyApprovedErrorMessage);
-
-
-            tour.Status = Status.Approved;
-            await repository.SaveChangesAsync();
-
-            response.StatusCode = HttpStatusCode.OK;
-            response.Result = this.MapTourToTourResponseDto(tour);
-
-            return response;
-        }
-
-
         public async Task<Tour> CreateAsync(TourCreateDTO model, string creatorId)
         {
             if (model == null) throw new ArgumentException();
@@ -225,20 +207,6 @@ namespace SelfGuidedTours.Core.Services
                 .ToListAsync();
         }
 
-        public async Task<ApiResponse> RejectTourAsync(int id)
-        {
-            var tour = await repository.GetByIdAsync<Tour>(id)
-                ?? throw new KeyNotFoundException(TourNotFoundErrorMessage);
-
-            if (tour.Status == Status.Declined) throw new InvalidOperationException(TourAlreadyRejectedErrorMessage);
-
-            tour.Status = Status.Declined;
-            await repository.SaveChangesAsync();
-
-            response.StatusCode = HttpStatusCode.OK;
-            response.Result = this.MapTourToTourResponseDto(tour);
-
-            return response;
-        }
+      
     }
 }
